# src/backend/doc/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.2.0",
    "title": "Caliopen HTTP/REST API"
  },
  "schemes": [
    "http"
  ],
  "host": "localhost:31415",
  "basePath": "/api",
  "paths": {
...
    "/v2/tags": {
      "get": {
        "description": "Returns tags visible to current user according to given parameters",
        "tags": [
          "tags"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Tags returned",
            "schema": {
              "type": "object",
              "properties": {
                "total": {
                  "type": "integer",
                  "format": "int32",
                  "description": "number of tags found for user for the given parameters"
                },
                "tags": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "date_insert": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "tag_id": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string",
                        "enum": [
                          "user",
                          "system"
                        ]
                      },
                      "name": {
                        "type": "string"
                      },
                      "importance_level": {
                        "type": "integer",
                        "format": "int32"
                      }
                    },
                    "required": [
                      "label"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "Create a new Tag for an user",
        "tags": [
          "tags"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "tag",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "importance_level": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "required": [
                "name"
              ],
              "additionalProperties": false
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "User tag creation completed",
            "schema": {
              "type": "object",
              "properties": {
                "location": {
                  "type": "string",
                  "description": "url to retrieve new tag's infos at /tags/{name}"
                }
              }
            }
          },
          "400": {
            "description": "malform request",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "name": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "name": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable entity. Parameters were valid but the server was unable to create the tag (probably because of a semantically erroneous document, like an empty tag)",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "name": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v2/tags/{tag_id}": {
      "get": {
        "description": "Retrieve tag infos",
        "tags": [
          "tags"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "parameters": [
          {
            "name": "tag_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful response with json object",
            "schema": {
              "type": "object",
              "properties": {
                "date_insert": {
                  "type": "string",
                  "format": "date-time"
                },
                "tag_id": {
                  "type": "string"
                },
                "type": {
                  "type": "string",
                  "enum": [
                    "user",
                    "system"
                  ]
                },
                "name": {
                  "type": "string"
                },
                "importance_level": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "required": [
                "label"
              ],
              "additionalProperties": false
            }
          },
          "401": {
            "description": "Unauthorized access",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "name": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "description": "update a tag",
        "tags": [
          "tags"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "tag_id",
            "in": "path",
            "type": "string",
            "required": true
          },
          {
            "name": "patch",
            "in": "body",
            "required": true,
            "description": "the patch to apply. See 'Caliopen Patch RFC' within /doc directory.",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "current_state": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    }
                  },
                  "additionalProperties": false,
                  "required": [
                    "name"
                  ]
                }
              },
              "additionalProperties": false,
              "required": [
                "name",
                "current_state"
              ]
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "204": {
            "description": "Update successful. No body is returned."
          },
          "400": {
            "description": "json payload malformed",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "name": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "name": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden patch. Server is refusing to apply the given patch's properties to this ressource",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "name": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "tag not found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "name": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "json is valid but patch conflicts with db's current state",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "name": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "json is valid but patch was semantically malformed or unprocessable",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "name": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "Delete a tag belonging to an user",
        "tags": [
          "tags"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "parameters": [
          {
            "name": "tag_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful deletion"
          },
          "401": {
            "description": "Unauthorized access",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "name": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "basicAuth": {
      "type": "basic",
      "description": "HTTP Basic Authentication. `Password` is the `access_token` return by /authentications and `Username` is the `user_id` returned by /authentications"
    }
  }
}